DOCS: 이전 구현 API 폴더 위치 변경 -> 테스트 용으로 남긴 이전 API 위치 바꾸기
