+++ src/main/java/spot/spot/domain/job/command/controller/ClientCommandController.java
@@ -47,11 +47,6 @@ public class ClientCommandController implements ClientCommandDocs {
         clientCommandService.requestWithdrawal(request);
     }
 
-    @PostMapping("/withdrawal/test")
-    public void requestWithdrawalTest(@RequestBody ChangeStatusClientRequest request) {
-        clientCommandService.requestWithdrawalTest(request);
-    }
-
     @PatchMapping("/confirm-or-reject")
     public void confirmOrRejectJob(@RequestBody YesOrNoWorkersRequest request) {
         clientCommandService.confirmOrRejectJob(request);
