+++ src/main/java/spot/spot/domain/job/v1/query/controller/_docs/WorkerQueryVersion1Docs.java
@@ -7,7 +7,7 @@ import org.springframework.data.domain.Slice;
 import org.springframework.web.bind.annotation.RequestParam;
 import spot.spot.domain.job.query.dto.response.NearByJobResponse;
 
-@Tag(name= "가. WORKER QUERY API VERSION 1", description = "<br/> 첫 구현 서비스 로직 테스트용 API")
+@Tag(name= "가. (V1) WORKER QUERY API ", description = "<br/> 해결사 R API (첫 구현)")
 public interface WorkerQueryVersion1Docs {
 
     @Operation(summary = "JPQL로 DB 쿼리로 거리 계산 후 리스트 출력 (default 위도 경도 현 사용자의 위도 경도)")
